gateway/requirements.txt
aiohttp==3.13.3
alembic==1.14.1
amqp==5.3.1
annotated-types==0.7.0
anyio==4.12.1
asyncpg==0.30.0
billiard==4.2.4
cachetools==5.5.2
celery==5.6.2
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
click==8.3.1
cryptography==46.0.5
distro==1.9.0
fastapi==0.115.6
google-api-core==2.29.0
google-api-python-client==2.159.0
google-auth-httplib2==0.3.0
google-auth==2.37.0
googleapis-common-protos==1.72.0
greenlet==3.3.1
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.2
httptools==0.7.1
httpx==0.28.1
idna==3.11
jiter==0.13.0
kombu==5.6.2
Mako==1.3.10
MarkupSafe==3.0.3
openai==1.59.7
packaging==26.0
prompt_toolkit==3.0.52
proto-plus==1.27.1
protobuf==6.33.5
psycopg2-binary==2.9.10
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
pydantic-settings==2.7.1
pydantic==2.10.4
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-telegram-bot==21.9
PyYAML==6.0.3
redis==7.1.1
requests==2.32.5
rsa==4.9.1
six==1.17.0
slack_sdk==3.34.0
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.41.3
tqdm==4.67.3
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
uritemplate==4.2.0
urllib3==2.6.3
uvicorn==0.34.0
uvloop==0.22.1
vine==5.1.0
watchfiles==1.1.1
wcwidth==0.6.0
websockets==16.0
